Break up large keychain source file
